refactor: extract ExportReceipt.swift from ExportReceiptStore.swift
Extracts the ExportReceipt Codable struct (49 lines) from ExportReceiptStore.swift
into its own file.

Byte-identical move; SHA-256 376ca901bfd0e767db7535f13fbe7e6e0df5a2293d054bb54aa8abd418cdf2a2
verified against origin/main:ExportReceiptStore.swift lines 3-51.

Reduces ExportReceiptStore.swift from 238 to 188 lines. The primary
ExportReceiptStore actor + TrackerExportFingerprint enum remain in the store file.

Tests: 667 (unchanged).
